+++ BombaProMaxApi/DTOs/BonLivraisonDetailsDto.cs
@@ -6,7 +6,7 @@ public class BonLivraisonDetailsDto
     public int BonLivraisonID { get; set; }
     public int? ProduitID { get; set; }
     public int? ServiceID { get; set; }
-    public int Quantite { get; set; }
+    public decimal Quantite { get; set; }
     public decimal PrixUnitaire { get; set; }
     public decimal MontantLigne { get; set; }
     public string? Description { get; set; }
@@ -22,7 +22,7 @@ public class CreateBonLivraisonDetailsDto
 {
     public int? ProduitID { get; set; }
     public int? ServiceID { get; set; }
-    public int Quantite { get; set; }
+    public decimal Quantite { get; set; }
     public decimal PrixUnitaire { get; set; }
     public string? Description { get; set; }
 }
