+++ BombaProMaxApi/DTOs/CreditTransactionDto.cs
@@ -8,7 +8,7 @@ public class CreditTransactionDto
     public int? ProduitID { get; set; }
     public int? ServiceID { get; set; }
     public decimal PrixTTC { get; set; }
-    public int Quantite { get; set; }
+    public decimal Quantite { get; set; }
     public decimal MontantTotal { get; set; }
     public DateTime DateCredit { get; set; }
     public bool Facture { get; set; }
